CDNIT-258; Atualização do tipo de dado do Contrato_id

+++ database/migrations/2026_08_09_160654_add_pmqa_and_contrato_id_to_modulos_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             }
 
             if (!Schema::hasColumn('modulos', 'contrato_id')) {
-                $table->unsignedBigInteger('contrato_id')->nullable()->after('pmqa');
+                $table->integer('contrato_id')->nullable()->after('pmqa');
                 $table->foreign('contrato_id')->references('id')->on('contratos')->nullOnDelete();
             }
         });
